Refactor TUI: Break up 3,097-line textual_browser.py monolith
Pure lift-and-shift refactoring to improve maintainability while preserving 100% identical functionality:

### Files Created:
- `emdx/ui/text_areas.py` (595 lines) - SelectionTextArea, VimEditTextArea
- `emdx/ui/inputs.py` (64 lines) - TitleInput
- `emdx/ui/document_viewer.py` (164 lines) - FullScreenView
- `emdx/ui/modals.py` (73 lines) - DeleteConfirmScreen
- `emdx/ui/main_browser.py` (2,385 lines) - MinimalDocumentBrowser + VimLineNumbers

### Result:
- **Before**: 1 monolithic file (3,097 lines)
- **After**: 6 focused files with clear separation of concerns
- **Zero functional changes** - pure code movement with imports
- **100% backward compatibility** - existing code continues to work
