+++ emdx/ui/modals.py
@@ -1,0 +1,74 @@
+#!/usr/bin/env python3
+"""
+Modal screens for EMDX TUI.
+"""
+
+from textual.app import ComposeResult
+from textual.containers import Grid
+from textual.screen import ModalScreen
+from textual.widgets import Button, Label
+
+
+class DeleteConfirmScreen(ModalScreen):
+    """Modal screen for delete confirmation."""
+
+    CSS = """
+    DeleteConfirmScreen {
+        align: center middle;
+    }
+
+    #dialog {
+        grid-size: 2;
+        grid-gutter: 1 2;
+        grid-rows: 1fr 3;
+        padding: 0 2;
+        width: 60;
+        height: 11;
+        border: thick $background 80%;
+        background: $surface;
+    }
+
+    #question {
+        column-span: 2;
+        height: 3;
+        content-align: center middle;
+        text-style: bold;
+    }
+
+    Button {
+        width: 100%;
+    }
+    """
+
+    BINDINGS = [
+        ("y", "confirm_delete", "Yes"),
+        ("n", "cancel", "No"),
+        ("escape", "cancel", "Cancel"),
+    ]
+
+    def __init__(self, doc_id: int, doc_title: str):
+        super().__init__()
+        self.doc_id = doc_id
+        self.doc_title = doc_title
+
+    def compose(self) -> ComposeResult:
+        with Grid(id="dialog"):
+            yield Label(
+                f'Delete document #{self.doc_id}?\n"{self.doc_title}"\n\n'
+                f"[dim]Press [bold]y[/bold] to delete, [bold]n[/bold] to cancel[/dim]",
+                id="question",
+            )
+            yield Button("Cancel (n)", variant="primary", id="cancel")
+            yield Button("Delete (y)", variant="error", id="delete")
+
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        if event.button.id == "delete":
+            self.dismiss(True)
+        else:
+            self.dismiss(False)
+
+    def action_confirm_delete(self) -> None:
+        self.dismiss(True)
+
+    def action_cancel(self) -> None:
+        self.dismiss(False)
